-Added the expansion caching optimization for 'expandables'. -Added edges set sharing to the children of 'expandables'. -Cycles are now detected on 'expandables' as well, instead of indirectly through their children. -Simplified the expansion code. -Optimized performance of 'expandables'. -Updated documentation related to this feature.
NOTE: This stuff is now ready to be ported to the Rascal version, as it fixes 'everything'. Additionally it also makes the implementation cubic in theory as well (that should satisfy the mathematicians among us ;-)). A three fold performance increase is to be expected (if I ever get round to actually do this).

Next on the list:
-Debugger features (to support runtime parse trace visualization).
-Expansion time filtering for priorities, associativity and disallowed nesting, instead of at reduction time (I have a brilliant plan :-)). This should change the worst-case behavior of most unambiguous grammars (after filtering) to O(N) (O(N^2) worst-case), instead of O(N^2) or O(N^3).